Move string_builtin_function class to a new file
The amount of code for these classes will grow as we support more and
more functions, so this will avoid making string_refinement_util grow
too much.
